Added Score Counter and added End Game Menu to Game Manager
So not sure what we decided in storing score, but i commented where you
can get the score. Also set up the end game menu, but I'm not sure why
it wont render once the game is won, I think it has something to do with
the image icon, but im not sure how to fix it.